+++ src/pages/home-page/index.ts
@@ -1,0 +1,2 @@
+export { default as HomePageLayout, metadata } from './ui/layout';
+export { default as HomePage } from './ui/home-page';
